Rename : PO-20 컴포넌트 이름 컨벤션 적용
